requirements.txt
pandas==2.0.0
numpy>=1.24.2
PyQt5>=5.15.9
scikit-learn>=1.3.0
ipython>=8.14.0
graphviz>=0.20.1
seaborn>=0.12.2
scikit-plot>=0.3.7
tensorflow>=2.13.0
configparser>=6.0.0
matplotlib
tensorrt
nvidia-tensorrt
umap-learn
scipy
lightgbm